Updated NetworkWrapper from other repo

+++ src/network_wrapper/training/scripts/generate_csv.py
@@ -24,8 +24,8 @@ def xml_to_csv(xml_dir, output_file):
             value = (root.find('filename').text,
                      int(root.find('size')[0].text),
                      int(root.find('size')[1].text), member[0].text,
-                     int(member[4][0].text), int(member[4][1].text),
-                     int(member[4][2].text), int(member[4][3].text))
+                     float(member[4][0].text), float(member[4][1].text),
+                     float(member[4][2].text), float(member[4][3].text))
             annotations.append(value)
 
     __list_to_csv(annotations, output_file)
